refac(crypto): extract `DuplexSpongeMode` from `sponge.h`
