+++ tachyon/crypto/hashes/sponge/sponge.h
@@ -9,12 +9,11 @@
 #include <numeric>
 #include <vector>
 
-#include "tachyon/base/buffer/copyable.h"
 #include "tachyon/base/logging.h"
+#include "tachyon/crypto/hashes/sponge/duplex_sponge_mode.h"
 #include "tachyon/math/finite_fields/finite_field_traits.h"
 
-namespace tachyon {
-namespace crypto {
+namespace tachyon::crypto {
 
 // Specifying the output field element size.
 class TACHYON_EXPORT FieldElementSize {
@@ -142,74 +141,6 @@ class FieldBasedCryptographicSponge : public CryptographicSponge<Derived> {
   }
 };
 
-// The mode structure for duplex sponge.
-struct TACHYON_EXPORT DuplexSpongeMode {
-  enum class Type {
-    // The sponge is currently absorbing data.
-    kAbsorbing,
-    // The sponge is currently squeezing data out.
-    kSqueezing,
-  };
-
-  constexpr DuplexSpongeMode() = default;
-
-  constexpr static DuplexSpongeMode Absorbing(size_t next_index = 0) {
-    return {Type::kAbsorbing, next_index};
-  }
-  constexpr static DuplexSpongeMode Squeezing(size_t next_index = 0) {
-    return {Type::kSqueezing, next_index};
-  }
-
-  Type type = Type::kAbsorbing;
-  // When |type| is |kAbsorbing|, it is interpreted as next position of the
-  // state to be XOR-ed when absorbing.
-  // When |type| is |kSqueezing|, it is interpreted as next position of the
-  // state to be outputted when squeezing.
-  size_t next_index = 0;
-
-  bool operator==(const DuplexSpongeMode& other) const {
-    return type == other.type && next_index == other.next_index;
-  }
-  bool operator!=(const DuplexSpongeMode& other) const {
-    return !operator==(other);
-  }
-
- private:
-  friend class base::Copyable<DuplexSpongeMode>;
-
-  constexpr DuplexSpongeMode(Type type, size_t next_index)
-      : type(type), next_index(next_index) {}
-};
-
-}  // namespace crypto
-
-namespace base {
-
-template <>
-class Copyable<crypto::DuplexSpongeMode> {
- public:
-  static bool WriteTo(const crypto::DuplexSpongeMode& mode, Buffer* buffer) {
-    return buffer->WriteMany(mode.type, mode.next_index);
-  }
-
-  static bool ReadFrom(const ReadOnlyBuffer& buffer,
-                       crypto::DuplexSpongeMode* mode) {
-    crypto::DuplexSpongeMode::Type type;
-    size_t next_index;
-    if (!buffer.ReadMany(&type, &next_index)) {
-      return false;
-    }
-
-    *mode = {type, next_index};
-    return true;
-  }
-
-  static size_t EstimateSize(const crypto::DuplexSpongeMode& mode) {
-    return base::EstimateSize(mode.type, mode.next_index);
-  }
-};
-
-}  // namespace base
-}  // namespace tachyon
+}  // namespace tachyon::crypto
 
 #endif  // TACHYON_CRYPTO_HASHES_SPONGE_SPONGE_H_
